Initial go at generating .ghc.environment files for projects
This code half works, but the environments it generates so far are not
quite correct. Also, ghc-8.0.1 doesn't seem to interpret them correctly.

We need to generate the package db stack properly, and select the right
packages to put into the env. We'll also want to do this after building.